[BUGFIX] Do not trust image dimensions from EXIF data
The image dimensions as stored within EXIF metadata should not be trusted
since they may not reflect the actual size of an image previously resized
while keeping the original metadata. As such, best is to let TYPO3 figure
the height/width of an image instead of extracting them.

It still makes sense though to extract this information when dealing with
e.g., PDF documents.

Change-Id: I0c1cc83cbca149cb0b4288ae35e7b1670eae285b
Resolves: #79213
Reviewed-on: https://review.typo3.org/51522
Tested-by: Rafal Brzeski <r.brzeski@gmail.com>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>