Add try/catch to processor calls
Because of some php 8.1 deprecations in e.g. mb_detect_encoding calls (used in PhpService::safeUtf8Encode()), where the first parameter must not be null, a processor call with $parameter[0]=null will throw an exception. To avoid this and other fails of processing, I added a try/catch. 
If the process fails, the origin value will be taken.