Feature/49 make php 8 compatible (#50)
* Fix some php 8 issues

* Fix php 8 issue

* Replace none php 7.4 compatible str_contains with strpos

* Revert array key change of dataMappingHook

* Add try/catch to processor calls

Because of some php 8.1 deprecations in e.g. mb_detect_encoding calls (used in PhpService::safeUtf8Encode()), where the first parameter must not be null, a processor call with $parameter[0]=null will throw an exception. To avoid this and other fails of processing, I added a try/catch. 
If the process fails, the origin value will be taken.

* Make safeUtf8Encode php 8 compatible

* Improve php 8 compatibility

* Make extractFloat php 8 compatible